Added in the missing sharing examples for AsyncFunSpec.
